cea_thermo_ascii_parsing no longer header only
