+++ src/Makefile.am
@@ -20,6 +20,7 @@ libantioch_la_SOURCES += parsing/src/ascii_parser.C
 libantioch_la_SOURCES += parsing/src/blottner_parsing.C
 libantioch_la_SOURCES += parsing/src/cea_mixture_parsing.C
 libantioch_la_SOURCES += parsing/src/cea_mixture_ascii_parsing.C
+libantioch_la_SOURCES += parsing/src/cea_thermo_ascii_parsing.C
 libantioch_la_SOURCES += parsing/src/chemkin_parser.C
 libantioch_la_SOURCES += parsing/src/nasa_mixture_parsing.C
 libantioch_la_SOURCES += parsing/src/parser_base.C
