+++ src/parsing/include/antioch/cea_thermo_ascii_parsing.h
@@ -31,13 +31,8 @@
 #ifndef ANTIOCH_CEA_THERMO_ASCII_PARSING_H
 #define ANTIOCH_CEA_THERMO_ASCII_PARSING_H
 
-// Antioch
-#include "antioch/chemical_mixture.h"
-#include "antioch/input_utils.h"
-
 // C++
-#include <iostream>
-#include <vector>
+#include <string>
 
 namespace Antioch
 {
@@ -48,71 +43,8 @@ namespace Antioch
   // New declarations
 
   template<class NumericType>
-  void read_cea_thermo_data_ascii( CEAThermodynamics<NumericType>& thermo, const std::string &filename );
-
-  /* ------------------------- Inline Functions -------------------------*/
-  template<class NumericType>
-  inline
-  void read_cea_thermo_data_ascii( CEAThermodynamics<NumericType>& thermo, const std::string &filename )
-  {
-    std::ifstream in(filename.c_str());
-    if(!in.is_open())
-    {
-      std::cerr << "ERROR: unable to load file " << filename << std::endl;
-      antioch_error();
-    }
-
-    skip_comment_lines(in, '#');
-
-    std::string name;
-    unsigned int n_int;
-    std::vector<NumericType> coeffs;
-    NumericType h_form, val;
-
-    const ChemicalMixture<NumericType>& chem_mixture = thermo.chemical_mixture();
-
-    while (in.good())
-      {
-	in >> name;   // Species Name
-	in >> n_int;  // Number of T intervals: [200-1000], [1000-6000], ([6000-20000])
-	in >> h_form; // Formation Enthalpy @ 298.15 K
-
-	coeffs.clear();
-	for (unsigned int interval=0; interval<n_int; interval++)
-	  {
-	    for (unsigned int n=0; n<10; n++)
-	      {
-		in >> val, coeffs.push_back(val);
-	      }
-	  }
-
-	// If we are still good, we have a valid set of thermodynamic
-	// data for this species. Otherwise, we read past end-of-file 
-	// in the section above
-	if (in.good())
-	  {
-	    // Check if this is a species we want.
-	    if( chem_mixture.species_name_map().find(name) !=
-		chem_mixture.species_name_map().end() )
-	      {
-		thermo.add_curve_fit(name, coeffs);
-	      }
-	  }
-      } // end while
-
-    in.close();
-    
-    // Make sure we actually populated everything
-    if( !thermo.check() )
-      {
-	std::cerr << "Error: CEA table not fully populated" << std::endl;
-	antioch_error();
-      }
-
-    return;
-  }
-
-
+  void read_cea_thermo_data_ascii( CEAThermodynamics<NumericType>& thermo,
+                                   const std::string &filename );
 
 } // end namespace Antioch
